+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update readme.
 
+### Fixed
+
+- Wrong Netlify badge.
+
 ## [0.0.5] - 2021-07-13
 
 ### Added