+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Track near point (as an addon of viewing distance tracking).
+- Track near point (as an addon to the viewing distance tracking).
 - Bot detection: `.bot` will return the name, category, and producer of the bot if the user agent is one, an empty string will be returned otherwise.
 - `.isFullscreen` getter to get the current window mode. `.getFullscreen()` now also records fullscreen mode status data.
 - `.version` getter on the demo page.