+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.7] - 2021-07-17
+
+No new feature updates in this release. Updated dependency packages and license.
+
 ## [0.0.6] - 2021-07-16
 
 ### Added
@@ -104,7 +108,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 The framework and some basic functions, e.g., screen size calibration. Released for integration testing.
 
-[unreleased]: https://github.com/EasyEyes/remote-calibrator/compare/v0.0.6...HEAD
+[unreleased]: https://github.com/EasyEyes/remote-calibrator/compare/v0.0.7...HEAD
+[0.0.7]: https://github.com/EasyEyes/remote-calibrator/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/EasyEyes/remote-calibrator/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/EasyEyes/remote-calibrator/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/EasyEyes/remote-calibrator/compare/v0.0.3...v0.0.4